refactor(catalog): relocate AliasBase from apps.core to apps.catalog
AliasBase is only used by catalog code (its 7 concrete subclasses, the
alias registry, resolve, and claims). Living in apps.core forced any
typing of its claim-control invariant to fight the layering rule that
core depends on nothing. Moving it into apps.catalog.models.base,
alongside CatalogModel, lets the next pass declare the parent_model
ClassVar as type[ClaimControlledModel] without inverting layers.

Also tightens the _alias_registry docstring (catalog-private; lives in
its own module to break a cycle between claims.py and resolve/) and
adds a follow-up note in ModelDrivenLinkability.md flagging the small
typing cleanup that becomes available once Location lands as a
LinkableModel — retype the entity registry as type[CatalogModel] and
delete the issubclass guard at config/api.py.
