+++ backend/apps/catalog/models/taxonomy.py
@@ -8,7 +8,6 @@ from django.db import models
 from django.db.models.functions import Lower
 
 from apps.core.models import (
-    AliasBase,
     MarkdownField,
     SluggedModel,
     TimeStampedModel,
@@ -18,7 +17,7 @@ from apps.core.models import (
 )
 from apps.core.validators import validate_no_mojibake
 
-from .base import CatalogModel
+from .base import AliasBase, CatalogModel
 
 if TYPE_CHECKING:
     from .machine_model import MachineModel
