+++ backend/apps/catalog/models/theme.py
@@ -8,7 +8,6 @@ from django.db import models
 from django.db.models.functions import Lower
 
 from apps.core.models import (
-    AliasBase,
     MarkdownField,
     SluggedModel,
     TimeStampedModel,
@@ -18,7 +17,7 @@ from apps.core.models import (
 )
 from apps.core.validators import validate_no_mojibake
 
-from .base import CatalogModel
+from .base import AliasBase, CatalogModel
 
 __all__ = ["MachineModelTheme", "Theme", "ThemeAlias"]
