+++ backend/apps/core/models.py
@@ -25,47 +25,6 @@ class TimeStampedModel(models.Model):
         abstract = True
 
 
-class AliasBase(TimeStampedModel):
-    """Abstract base for alias lookup models.
-
-    Alias values are stored and compared in lowercase (matching the
-    UniqueConstraint(Lower("value")) that every subclass must define).
-    Claims live on the *parent* object, not on the alias row itself.
-
-    Subclasses must add:
-    - A ForeignKey to the parent model (named after the parent, related_name="aliases")
-    - A UniqueConstraint on Lower("value") with a table-specific name
-    - ``alias_claim_field``: the claim namespace on the parent that carries
-      alias values (e.g. ``"theme_alias"``). Enforced at class creation
-      via ``__init_subclass__``; read by ``discover_alias_types``.
-    """
-
-    alias_claim_field: ClassVar[str]
-
-    value = models.CharField(max_length=200)
-
-    class Meta(TimeStampedModel.Meta):
-        abstract = True
-        ordering = ["value"]
-
-    def __init_subclass__(cls, **kwargs: object) -> None:
-        # NB: we can't gate on ``cls._meta.abstract`` here — Django's ModelBase
-        # runs ``__init_subclass__`` with ``abstract`` still inherited as True
-        # from the parent, then rewrites it to False for concrete subclasses
-        # later. So this check runs for every AliasBase subclass, concrete or
-        # not. That's fine: any abstract intermediate can just declare
-        # ``alias_claim_field`` for its concrete descendants to inherit.
-        super().__init_subclass__(**kwargs)
-        if not getattr(cls, "alias_claim_field", None):
-            raise TypeError(
-                f"{cls.__name__} must declare a non-empty `alias_claim_field` "
-                'class attr (e.g. `alias_claim_field = "theme_alias"`)'
-            )
-
-    def __str__(self) -> str:
-        return self.value
-
-
 def field_not_blank(field_name: str) -> models.CheckConstraint:
     """CHECK constraint: field != ''. Use in concrete model Meta.constraints."""
     return models.CheckConstraint(
